# new_swagger.yml
openapi: '3.0.2'
info:
  title: Tribe App
  version: '1.0'
paths:
...
  /users/user_id/image/:
    post:
      tags:
      - Profile image
      summary: Upload Profile Image
      responses:
        '201':
          description: OK
        '401':
          description: Invalid Usern
        '400':
          description: Missing image
        '402':
          description: Invalid File Type 
    get:
      tags:
      - Profile image
      summary: Retrive Image
      responses:
        '201':
          description: OK
        '401':
          description: Invalid User
    delete:
      tags:
      - Profile image
      summary: Delete Profile Image
      responses:
        '204':
          description: OK
        '401':
          description: Invalid User